requirements.txt
psycopg2
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
passlib[argon2]
python-multipart
pydantic[email]

python-jose[cryptography]